requirements.txt
3==0.0.0
aniso8601==7.0.0
asgiref==3.5.0
attrs==21.4.0
autopep8==1.6.0
black==21.12b0
cfgv==3.3.1
click==8.0.3
coverage==6.3
distlib==0.3.4
Django==4.0.1
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.13.1
Faker==11.3.0
filelock==3.4.2
flake8==4.0.1
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
identify==2.4.5
importlib-metadata==4.10.1
iniconfig==1.1.1
Markdown==3.3.6
mccabe==0.6.1
mixer==7.2.0
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.1
pluggy==1.0.0
pre-commit==2.17.0
promise==2.3
psycopg2==2.9.3
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.7
pytest==6.2.5
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tomli==1.2.3
typing_extensions==4.0.1
virtualenv==20.13.0
zipp==3.7.0
